+++ paper/paper.bib
@@ -2,6 +2,7 @@
   author    = {Christopher Rackauckas and Yingbo Ma and Julius Martensen and Collin Warner and Kirill Zubov and
                Rohit Supekar and Dominic Skinner and Ali Jasim Ramadhan},
   title     = {Universal Differential Equations for Scientific Machine Learning},
+  doi 		= {https://doi.org/10.21203/rs.3.rs-55125/v1},
   journal   = {CoRR},
   volume    = {abs/2001.04385},
   year      = {2020}
@@ -31,6 +32,7 @@
 @article{Drake2021,
 	title = {{A simple model for assessing climate control trade-offs and responding to unanticipated climate outcomes}},
 	author = {Henri F. Drake and Ronald L. Rivest and Alan Edelman and John Deutch},
+	doi = {10.1088/1748-9326/ac243e},
     journal = {Environmental Research Letters},
     volume = {16},
     number = {10},
@@ -97,6 +99,7 @@
 	title = {The Earth4All model of human wellbeing on a finite planet towards 2100},
 	url = {https://www.clubofrome.org/wp-content/uploads/2022/09/220916_E4A_technical-note.pdf},
 	author = {Jorgen Randers and David Collste},
+	doi = {10.31223/X50W8D}
 	keywords = {System dynamics, Integrated assessment model, e4a},
 	pages = {66},
 	year = {2022},
@@ -124,6 +127,7 @@
 
 @article{Moore2018,
 	title = {Mimi-{PAGE}, an open-source implementation of the {PAGE09} integrated assessment model},
+	doi = {10.1038/sdata.2018.187},
 	volume = {5},
 	number = {1},
 	journal = {Scientific Data},
@@ -135,6 +139,7 @@
 @article{Lubin2023,
     author = {Miles Lubin and Oscar Dowson and Joaquim Dias Garcia and Joey Huchette and Beno{\^i}t Legat and Juan Pablo Vielma},
     title = {JuMP 1.0: Recent improvements to a modeling language for mathematical optimization},
+	doi = {10.48550/arXiv.2206.03866}
     journal = {Mathematical Programming Computation},
     year = {2023},
     note = {In press.}
@@ -143,6 +148,7 @@
 @misc{Ma2021,
       title={ModelingToolkit: A Composable Graph Transformation System For Equation-Based Modeling},
       author={Yingbo Ma and Shashi Gowda and Ranjan Anantharaman and Chris Laughman and Viral Shah and Chris Rackauckas},
+	  doi= {10.48550/arXiv.2103.05244}
       year={2021},
       eprint={2103.05244},
       archivePrefix={arXiv},
@@ -151,6 +157,7 @@
 
 @article{Rackauckas2017,
   title={Differentialequations.jl--a performant and feature-rich ecosystem for solving differential equations in julia},
+  doi= {10.5334/jors.151},
   author={Rackauckas, Christopher and Nie, Qing},
   journal={Journal of Open Research Software},
   volume={5},
@@ -174,6 +181,7 @@
 @article{Dixit2022,
   title={GlobalSensitivity. jl: Performant and Parallel Global Sensitivity Analysis with Julia},
   author={Vaibhav Kumar Dixit and Christopher Rackauckas},
+  doi = {10.21105/joss.04561},
   journal={Journal of Open Source Software},
   volume={7},
   number={76},
@@ -184,6 +192,7 @@
 
 @article{Parker2002,
 	title = {Progress in integrated assessment and modelling},
+	doi = {10.1016/S1364-8152(01)00059-7},
 	volume = {17},
 	number = {3},
 	journal = {Environmental Modelling \& Software},
@@ -191,3 +200,5 @@
 	year = {2002},
 	pages = {209--217},
 }
+
+
