+++ paper/paper.bib
@@ -99,7 +99,7 @@
 	title = {The Earth4All model of human wellbeing on a finite planet towards 2100},
 	url = {https://www.clubofrome.org/wp-content/uploads/2022/09/220916_E4A_technical-note.pdf},
 	author = {Jorgen Randers and David Collste},
-	doi = {10.31223/X50W8D}
+	doi = {10.31223/X50W8D},
 	keywords = {System dynamics, Integrated assessment model, e4a},
 	pages = {66},
 	year = {2022},
@@ -139,7 +139,7 @@
 @article{Lubin2023,
     author = {Miles Lubin and Oscar Dowson and Joaquim Dias Garcia and Joey Huchette and Beno{\^i}t Legat and Juan Pablo Vielma},
     title = {JuMP 1.0: Recent improvements to a modeling language for mathematical optimization},
-	doi = {10.48550/arXiv.2206.03866}
+	doi = {10.48550/arXiv.2206.03866},
     journal = {Mathematical Programming Computation},
     year = {2023},
     note = {In press.}
@@ -148,7 +148,7 @@
 @misc{Ma2021,
       title={ModelingToolkit: A Composable Graph Transformation System For Equation-Based Modeling},
       author={Yingbo Ma and Shashi Gowda and Ranjan Anantharaman and Chris Laughman and Viral Shah and Chris Rackauckas},
-	  doi= {10.48550/arXiv.2103.05244}
+	  doi= {10.48550/arXiv.2103.05244},
       year={2021},
       eprint={2103.05244},
       archivePrefix={arXiv},
