+++ paper/paper.bib
@@ -2,7 +2,7 @@
   author    = {Christopher Rackauckas and Yingbo Ma and Julius Martensen and Collin Warner and Kirill Zubov and
                Rohit Supekar and Dominic Skinner and Ali Jasim Ramadhan},
   title     = {Universal Differential Equations for Scientific Machine Learning},
-  doi 		= {https://doi.org/10.21203/rs.3.rs-55125/v1},
+  doi 		= {10.21203/rs.3.rs-55125/v1},
   journal   = {CoRR},
   volume    = {abs/2001.04385},
   year      = {2020}
