### 2023-04-18 fc85871
Fix one typo in tutorial
