### 2023-04-18 c35caa0
Partially fix connection equations in alternative scenarios
