+++ src/World3/world3/plots.jl
@@ -18,13 +18,13 @@ function _variables_7()
     @named dr = Pop4.death_rate()
 
     variables = [
-        (nr.nrfr,  0, 1,    "nrfr"),
-        (is.iopc,  0, 1000, "iopc"),
-        (ld.fpc,   0, 1000, "fpc"),
-        (pop.pop,  0, 16e9, "pop"),
-        (pp.ppolx, 0, 32,   "ppolx"),
-        (br.cbr,   0, 50,   "cbr"),
-        (dr.cdr,   0, 50,   "cdr"),
+        (nr.nrfr, 0, 1, "nrfr"),
+        (is.iopc, 0, 1000, "iopc"),
+        (ld.fpc, 0, 1000, "fpc"),
+        (pop.pop, 0, 16e9, "pop"),
+        (pp.ppolx, 0, 32, "ppolx"),
+        (br.cbr, 0, 50, "cbr"),
+        (dr.cdr, 0, 50, "cdr"),
     ]
 
     return variables
@@ -40,13 +40,13 @@ function _variables_20()
     @named dr = Pop4.death_rate()
 
     variables = [
-        (nr.nrfr,  0, 1,    "nrfr"),
-        (is.iopc,  0, 2000, "iopc"),
-        (ld.fpc,   0, 1000, "fpc"),
-        (pop.pop,  0, 16e9, "pop"),
-        (pp.ppolx, 0, 32,   "ppolx"),
-        (br.cbr,   0, 50,   "cbr"),
-        (dr.cdr,   0, 50,   "cdr"),
+        (nr.nrfr, 0, 1, "nrfr"),
+        (is.iopc, 0, 2000, "iopc"),
+        (ld.fpc, 0, 1000, "fpc"),
+        (pop.pop, 0, 16e9, "pop"),
+        (pp.ppolx, 0, 32, "ppolx"),
+        (br.cbr, 0, 50, "cbr"),
+        (dr.cdr, 0, 50, "cdr"),
     ]
 
     return variables
@@ -65,16 +65,16 @@ function fig_2(; kwargs...)
     @named dr = Pop4.death_rate()
 
     variables = [
-        (pop.pop, 0,    4e9,  "pop"),
-        (br.cbr,  0,    50,   "cbr"),
-        (dr.cdr,  0,    50,   "cdr"),
-        (dr.le,   0,    60,   "le"),
-        (dr.lmf,  0.75, 1.75, "lmf"),
-        (dr.lmp,  0.75, 1.75, "lmp"),
+        (pop.pop, 0, 4e9, "pop"),
+        (br.cbr, 0, 50, "cbr"),
+        (dr.cdr, 0, 50, "cdr"),
+        (dr.le, 0, 60, "le"),
+        (dr.lmf, 0.75, 1.75, "lmf"),
+        (dr.lmp, 0.75, 1.75, "lmp"),
         (dr.lmhs, 0.75, 1.75, "lmhs"),
-        (dr.lmc,  0.75, 1.75, "lmc"),
-        (br.tf,   0,    8,    "tf"),
-        (br.dtf,  0,    8,    "dtf"),
+        (dr.lmc, 0.75, 1.75, "lmc"),
+        (br.tf, 0, 8, "tf"),
+        (br.dtf, 0, 8, "dtf"),
     ]
 
     return plotvariables(historicalrunsolution(), (t, 1900, 1970), variables; title="Fig. 7.2", kwargs...)
@@ -98,13 +98,13 @@ function fig_3(; kwargs...)
     @named se = SupplementaryEquations.supplementary_equations()
 
     variables = [
-        (is.ic,   0, 4e12, "ic"),
-        (is.io,   0, 4e12, "io"),
-        (is.iopc, 0, 400,  "iopc"),
[…]
