### 2023-04-18 220b375
Revert "Partially fix connection equations in alternative scenarios"
This reverts commit c35caa0218e7499ae3fd9691095c6c08f822147d.
