### 2023-04-21 eadcd0b
Add 4 connected sector of E4A
