+++ src/Earth4All/finance/initialisations.jl
@@ -1,0 +1,10 @@
+_inits = Dict{Symbol,Float64}(
+    :CCSD =>  0.15 + 0.005 + 0.02,
+    :ELTI => 0.02,
+    :PI => 0.02,
+    :PU => 0.0326951,
+    :CBSR => 0.02,
+)
+
+
+getinitialisations() = copy(_inits)
