+++ src/Earth4All/Wellbeing.jl
@@ -1,0 +1,13 @@
+module Wellbeing
+
+using WorldDynamics
+using ModelingToolkit
+
+include("wellbeing/tables.jl")
+include("wellbeing/parameters.jl")
+include("wellbeing/initialisations.jl")
+include("wellbeing/subsystems.jl")
+include("wellbeing/scenarios.jl")
+include("wellbeing/plots.jl")
+
+end
