+++ src/Earth4All/Public.jl
@@ -1,0 +1,13 @@
+module Public
+
+using WorldDynamics
+using ModelingToolkit
+
+include("public/tables.jl")
+include("public/parameters.jl")
+include("public/initialisations.jl")
+include("public/subsystems.jl")
+include("public/scenarios.jl")
+include("public/plots.jl")
+
+end

[…]
