### 2023-05-16 da4ad95
Remove not necessary tyoe
Co-authored-by: Emanuele Natale <natema@users.noreply.github.com>
