+++ src/solvesystems.jl
@@ -39,7 +39,7 @@ function variable_connections(systems::Vector{ODESystem})
     @variables t
     @named _model = ODESystem([], t)
     @named model = ModelingToolkit.compose(_model, systems)
-    connection_eqs::Set{Equation} = Set{Equation}()
+    connection_eqs = Set{Equation}()
     var2sys = Dict{String,String}()
     var2fullvar = Dict()
     g = variable_dependencies(model)
