+++ src/solvesystems.jl
@@ -44,7 +44,7 @@ function variable_connections(systems::Vector{ODESystem})
     var2fullvar = Dict()
     g = variable_dependencies(model)
     al = g.fadjlist
-    for u in 1:lastindex(al)
+    for u in eachindex(al)
         @assert length(al[u]) <= 1 "Error in the variable dependencies binary graph: one node has more than one neighbor"
         if (length(al[u]) > 0)
             s, v = split(string(states(model)[u]), "₊")
