### 2023-06-01 #5365 @fraukewiese
@gidden @khaeru @r-barnes
– would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
