### 2023-08-07 #5365 @fraukewiese
@ChrisRackauckas @SichengHe @StanczakDominik
– would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
