### 2023-08-21 #5772 @editorialbot
[REVIEW]: WorldDynamics.jl: A Julia Package for Developing and Simulating Integrated Assessment Models
**Submitting author:** @aurorarossi (<a href="http://orcid.org/0009-0001-3302-0468">Aurora Rossi</a>)
**Repository:** https://github.com/worlddynamics/WorldDynamics.jl
**Branch with paper.md** (empty if default branch): JOSS-paper
**Version:** v0.4.4
