### 2023-09-08 worlddynamics/WorldDynamics.jl#197 @ranocha
JOSS Review
# JOSS Review

Xref https://github.com/openjournals/joss-reviews/issues/5772

- [x] Contribution and authorship: Does the full list of paper authors seem appropriate and complete?
      - There is a description how to cite the package at https://github.com/worlddynamics/WorldDynamics.jl#how-to-cite-this-work. However, the author list there does not coincide with the JOSS author list.
      - Moreover, it looks like an additional person (`@universmile`) has made non-trivial contributions - at least based on the GitHub commit list and compared to other authors.
- [x] There are no installation instructions in the documentation. Since there is something in the README.md, I suggest to include it in the online docs, too.
- [x] Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers.
     I can only partially reproduce the results. First, it needs some experimentation to find out that "figure 7-7" or "Fig. 7.7" is reproduced by `World3.fig_7()`. This could be stated more explicitly. Next, I can't overlay the original plot - but I also don't have access to the book and putting it only may be complicated due to copyright restrictions.
- [x] The docs of many functions such as https://worlddynamics.github.io/WorldDynamics.jl/stable/source/#WorldDynamics.World1.fig_1-Tuple{} do not tell me how to call these functions - there is no function/method signature.
- [x] There seem to be some tests that can be run locally. However, there is no description of this in the docs. Moreover, there are no automated CI tests.
- [x] There are no community guidelines.
- [x] References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)?
      - Julia itself should be cited.
      - The paper mentions official implementations on Vensim and Stella without any references.
