### 2023-09-12 7a48f0a
Improve references to Stella etc
