+++ paper/paper.bib
@@ -189,7 +189,6 @@
   year={2022}
 }
 
-
 @article{Parker2002,
 	title = {Progress in integrated assessment and modelling},
 	doi = {10.1016/S1364-8152(01)00059-7},
@@ -201,4 +200,36 @@
 	pages = {209--217},
 }
 
+@misc{World3Vensim,
+  title = {{{WORLD3-03}} - {{MetaSD}}},
+  author = {Tom Fiddaman},
+  year = {2010},
+  month = apr,
+  howpublished = {https://web.archive.org/web/20230205104440/https://metasd.com/2010/04/world3-03/}
+}
 
+@inproceedings{World3Modelica,
+  title = {World3 in {{Modelica}}: Creating System Dynamics Models in the {{Modelica}} Framework},
+  shorttitle = {World3 in {{Modelica}}},
+  booktitle = {Proceedings of the 6th {{International Modelica Conference}}},
+  author = {Cellier, Fran{\c c}ois E.},
+  year = {2008},
+  volume = {2},
+  pages = {393--400},
+  publisher = {{The Modelica Association Bielefeld}}
+}
+
+
+@article{World3Simulink,
+  title = {Application of Fuzzy Tools to the Automatic Analysis of System Dynamics Models: An Example of {{World3}}},
+  shorttitle = {Application of Fuzzy Tools to the Automatic Analysis of System Dynamics Models},
+  author = {Mediavilla, Margarita and Miguel, Luis J. and Retortillo, Pedro and {de Castro}, Carlos and Bl{\'a}zquez, L. Felipe},
+  year = {2014},
+  month = jan,
+  journal = {IFAC Proceedings Volumes},
+  series = {19th {{IFAC World Congress}}},
+  volume = {47},
+  number = {3},
+  pages = {8983--8988},
+  doi = {10.3182/20140824-6-ZA-1003.01654}
+}
