### 2023-09-15 a35ae82
Fix world1 fig docstrings
