### 2023-09-15 8b40915
Fix world3 agr fig docstrings
