+++ src/World3/agriculture/plots.jl
@@ -70,67 +70,82 @@ end
 
 
 """
-    Reproduce Fig 4.69a. The original figure is presented in Chapter 4 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_69a(; kwargs...)
 
-    Caption: Run 4-1: historical run.
-    The behavior of land yields and food production.
+Reproduce Fig 4.69a. The original figure is presented in Chapter 4 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 4-1: historical run.
+The behavior of land yields and food production.
 """
 fig_69a(; kwargs...) = plotvariables(historicalrunsolution(), (t, 1900, 1970), _variables_a(); title="Fig. 4.69a", kwargs...)
 
 """
-    Reproduce Fig 4.69b. The original figure is presented in Chapter 4 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_69b(; kwargs...)
 
-    Caption: Run 4-1: historical run.
-    The behavior of arable land.
+Reproduce Fig 4.69b. The original figure is presented in Chapter 4 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 4-1: historical run.
+The behavior of arable land.
 """
 fig_69b(; kwargs...) = plotvariables(historicalrunsolution(), (t, 1900, 1970), _variables_b(); title="Fig. 4.69b", kwargs...)
 
 """
-    Reproduce Fig 4.69c. The original figure is presented in Chapter 4 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_69c(; kwargs...)
 
-    Caption: Run 4-1: historical run.
-    The allocation mechanism.
+Reproduce Fig 4.69c. The original figure is presented in Chapter 4 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 4-1: historical run.
+The allocation mechanism.
 """
 fig_69c(; kwargs...) = plotvariables(historicalrunsolution(), (t, 1900, 1970), _variables_c(); title="Fig. 4.69c", kwargs...)
 
 """
-    Reproduce Fig 4.69d. The original figure is presented in Chapter 4 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_69d(; kwargs...)
 
-    Caption: Run 4-1: historical run.
-    The behavior of land fertility.
+Reproduce Fig 4.69d. The original figure is presented in Chapter 4 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 4-1: historical run.
+The behavior of land fertility.
 """
 fig_69d(; kwargs...) = plotvariables(historicalrunsolution(), (t, 1900, 1970), _variables_d(); title="Fig. 4.69d", kwargs...)
 
 
 """
-    Reproduce Fig 4.70a. The original figure is presented in Chapter 4 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_70a(; kwargs...)
 
-    Caption: Run 4-2: standard run.
-    The behavior of land yields and food production.
+Reproduce Fig 4.70a. The original figure is presented in Chapter 4 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 4-2: standard run.
+The behavior of land yields and food production.
 """
 fig_70a(; kwargs...) = plotvariables(historicalrunsolution(), (t, 1900, 2100), _variables_a(); title="Fig. 4.70a", kwargs...)
 
 """
[…]
