### 2023-09-15 33c8140
Fix world3 capital
