+++ src/World3/capital/plots.jl
@@ -75,17 +75,21 @@ end
 
 
 """
-    Reproduce Fig 3.37. The original figure is presented in Chapter 3 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_37(; kwargs...)
 
-    Caption: Run 3-1: standard run of the capital sector with exogenous inputs.
+Reproduce Fig 3.37. The original figure is presented in Chapter 3 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 3-1: standard run of the capital sector with exogenous inputs.
 """
 fig_37(; kwargs...) = plotvariables(historicalrunsolution(), (t, 1900, 2000), _variables_37(); title="Fig. 3.37", kwargs...)
 
 
 """
-    Reproduce Fig 3.38. The original figure is presented in Chapter 3 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_38(; kwargs...)
 
-    Caption: Run 3-2: behavior of the capital sector when the average lifetime of industrial capital is increased from 14 to 21 years with standard inputs.
+Reproduce Fig 3.38. The original figure is presented in Chapter 3 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 3-2: behavior of the capital sector when the average lifetime of industrial capital is increased from 14 to 21 years with standard inputs.
 """
 function fig_38(; kwargs...)
     parameters_3_38 = getparameters()
@@ -98,9 +102,11 @@ function fig_38(; kwargs...)
 end
 
 """
-    Reproduce Fig 3.39. The original figure is presented in Chapter 3 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_39(; kwargs...)
 
-    Caption: Run3-3: behaviorofthecapital sector when the capital-output ratio is decreased from 3 to 2 years with standard inputs.
+Reproduce Fig 3.39. The original figure is presented in Chapter 3 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run3-3: behaviorofthecapital sector when the capital-output ratio is decreased from 3 to 2 years with standard inputs.
 """
 function fig_39(; kwargs...)
     parameters_3_39 = getparameters()
@@ -113,10 +119,12 @@ function fig_39(; kwargs...)
 end
 
 """
-    Reproduce Fig 3.40. The original figure is presented in Chapter 3 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_40(; kwargs...)
 
-    Caption: Run 3-4: behavior of the capital sector when the industrial capital-output ratio is increased from 3 to 4 years with standard inputs.
-    Note: Scales for IOPC, SOPC, and IO have been changed from their normal values.
+Reproduce Fig 3.40. The original figure is presented in Chapter 3 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 3-4: behavior of the capital sector when the industrial capital-output ratio is increased from 3 to 4 years with standard inputs.
+Note: Scales for IOPC, SOPC, and IO have been changed from their normal values.
 """
 function fig_40(; kwargs...)
     parameters_3_40 = getparameters()
@@ -129,10 +137,12 @@ function fig_40(; kwargs...)
 end
 
 """
-    Reproduce Fig 3.41. The original figure is presented in Chapter 3 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_41(; kwargs...)
 
[…]
