### 2023-09-15 10a5553
Fix world3 pollution
