+++ src/World3/pollution/plots.jl
@@ -24,9 +24,11 @@ end
 
 
 """
-    Reproduce Fig 6.26. The original figure is presented in Chapter 6 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_26(; kwargs...)
 
-    Caption: Run 6-1: behavior of the pollution sector in response to a pulse input in persistent pollution generation in 1920.
+Reproduce Fig 6.26. The original figure is presented in Chapter 6 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 6-1: behavior of the pollution sector in response to a pulse input in persistent pollution generation in 1920.
 """
 function fig_26(; kwargs...)
     @named pp = persistent_pollution()
@@ -57,9 +59,11 @@ function fig_26(; kwargs...)
 end
 
 """
-    Reproduce Fig 6.27. The original figure is presented in Chapter 6 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_27(; kwargs...)
 
-    Caption: Run 6-2: behavior of the pollution sector in response to a step increase and decrease in persistent pollution generation.
+Reproduce Fig 6.27. The original figure is presented in Chapter 6 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 6-2: behavior of the pollution sector in response to a step increase and decrease in persistent pollution generation.
 """
 function fig_27(; kwargs...)
     @named pp = persistent_pollution()
@@ -90,9 +94,11 @@ function fig_27(; kwargs...)
 end
 
 """
-    Reproduce Fig 6.28. The original figure is presented in Chapter 6 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_28(; kwargs...)
 
-    Caption: Inputs to Run 6-3, the historical run of the pollution sector.
+Reproduce Fig 6.28. The original figure is presented in Chapter 6 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Inputs to Run 6-3, the historical run of the pollution sector.
 """
 function fig_28(; kwargs...)
     @named nr = non_renewable()
@@ -112,9 +118,11 @@ function fig_28(; kwargs...)
 end
 
 """
-    Reproduce Fig 6.29. The original figure is presented in Chapter 6 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_29(; kwargs...)
 
-    Caption: Run 6-3: historical run of the pollution sector.
+Reproduce Fig 6.29. The original figure is presented in Chapter 6 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 6-3: historical run of the pollution sector.
 """
 function fig_29(; kwargs...)
     @named pp = persistent_pollution()
@@ -134,9 +142,11 @@ function fig_29(; kwargs...)
 end
 
 """
-    Reproduce Fig 6.30. The original figure is presented in Chapter 6 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_30(; kwargs...)
 
-    Caption: Inputs to Run 6-4 of the pollution sector when continued material growth is assumed.
+Reproduce Fig 6.30. The original figure is presented in Chapter 6 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Inputs to Run 6-4 of the pollution sector when continued material growth is assumed.
 """
 function fig_30(; kwargs...)
     @named nr = non_renewable()
@@ -157,16 +167,20 @@ end
 
 
 """
[…]
