+++ src/World3/population/pop1/plots.jl
@@ -62,30 +62,38 @@ end
 
 
 """
-    Reproduce Fig 2.84. The original figure is presented in Chapter 2 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_84(; kwargs...)
 
-    Caption: Run 2-1: historical behavior, 1900-1975.
+Reproduce Fig 2.84. The original figure is presented in Chapter 2 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 2-1: historical behavior, 1900-1975.
 """
 fig_84(; kwargs...) = plotvariables(historicalrunsolution(), (t, 1900, 1970), _variables_a(); title="Fig. 2.84a", kwargs...)
 
 """
-    Reproduce Fig 2.85. The original figure is presented in Chapter 2 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_85(; kwargs...)
 
-    Caption: Run 2—2: historical behavior, 1900—1975, mortality variables.
+Reproduce Fig 2.85. The original figure is presented in Chapter 2 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 2—2: historical behavior, 1900—1975, mortality variables.
 """
 fig_85(; kwargs...) = plotvariables(historicalrunsolution(), (t, 1900, 1970), _variables_b(); title="Fig. 2.85a", kwargs...)
 
 """
-    Reproduce Fig 2.86. The original figure is presented in Chapter 2 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_86(; kwargs...)
 
-    Caption: Run 2-3: historical behavior, 1900-1975, fertility variables.
+Reproduce Fig 2.86. The original figure is presented in Chapter 2 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 2-3: historical behavior, 1900-1975, fertility variables.
 """
 fig_86(; kwargs...) = plotvariables(historicalrunsolution(), (t, 1900, 1970), _variables_c(); title="Fig. 2.86a", kwargs...)
 
 """
-    Reproduce Fig 2.87. The original figure is presented in Chapter 2 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_87(; kwargs...)
 
-    Caption: Run 2-4: constant low income.
+Reproduce Fig 2.87. The original figure is presented in Chapter 2 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 2-4: constant low income.
 """
 function fig_87(; kwargs...)
     parameters_2_87 = getparameters()
@@ -99,9 +107,11 @@ function fig_87(; kwargs...)
 end
 
 """
-    Reproduce Fig 2.88. The original figure is presented in Chapter 2 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_88(; kwargs...)
 
-    Caption: Run 2-5: constant high income.
+Reproduce Fig 2.88. The original figure is presented in Chapter 2 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 2-5: constant high income.
 """
 function fig_88(; kwargs...)
     parameters_2_88 = getparameters()
@@ -118,9 +128,11 @@ function fig_88(; kwargs...)
 end
 
 """
-    Reproduce Fig 2.89. The original figure is presented in Chapter 2 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_89(; kwargs...)
 
-    Caption: Run 2-6: constant low income, improved health care.
[…]
