+++ src/World3/world3/plots.jl
@@ -53,11 +53,12 @@ function _variables_20()
 end
 
 """
-    Reproduce Fig 7.2. The original figure is presented in Chapter 7 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_2(; kwargs...)
 
-    Caption: Run 7-1: population sector behavior, 1900-1970.
-    Population POP increases over time at an average growth rate of 1.2 percent per year. Both the birth rate CBR and the death rate CDR decrease over the period, the former largely because of a lower desired total
-    fertility DTF, and the latter primarily as a result of increased health services LMHS. Both trends occur as a result of industrialization.
+Reproduce Fig 7.2. The original figure is presented in Chapter 7 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 7-1: population sector behavior, 1900-1970.
+Population POP increases over time at an average growth rate of 1.2 percent per year. Both the birth rate CBR and the death rate CDR decrease over the period, the former largely because of a lower desired total fertility DTF, and the latter primarily as a result of increased health services LMHS. Both trends occur as a result of industrialization.
 """
 function fig_2(; kwargs...)
     @named pop = Pop4.population()
@@ -81,10 +82,12 @@ function fig_2(; kwargs...)
 end
 
 """
-    Reproduce Fig 7.3. The original figure is presented in Chapter 7 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_3(; kwargs...)
 
-    Caption: Run 7-2: capital sector behavior, 1900-1970.
-    Industrial capital IC grows exponentially, causing industrial output IO to
+Reproduce Fig 7.3. The original figure is presented in Chapter 7 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 7-2: capital sector behavior, 1900-1970.
+Industrial capital IC grows exponentially, causing industrial output IO to
     grow. Since their growth rate is greater than that of population, industrial
     output per capita IOPC also grows over the period, as do service output
     per capita SOPC and food per capita (not graphed). As development proceeds, (1) the fraction of output in agriculture FOA declines, (2) FOA is
@@ -111,14 +114,16 @@ function fig_3(; kwargs...)
 end
 
 """
-    Reproduce Fig 7.4. The original figure is presented in Chapter 7 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+    fig_4(; kwargs...)
 
-    Caption: Run 7-3: agriculture sector behavior, 1900-1970.
-    Increases in arable land AL and land yields LY cause a rise in food
-    production over the historical period. The increase in land yields is
-    primarily attributable to greater agricultural inputs per hectare AIPH (fertilizers, pesticides), for the land fertility LFERT remains nearly constant.
-    Food per capita FPC also grows during the 70-year period but at a much
-    slower rate than total food F, since the population is also increasing.
+Reproduce Fig 7.4. The original figure is presented in Chapter 7 of [DGFW](https://archive.org/details/dynamicsofgrowth0000unse).
+
+Caption: Run 7-3: agriculture sector behavior, 1900-1970.
+Increases in arable land AL and land yields LY cause a rise in food
+production over the historical period. The increase in land yields is
+primarily attributable to greater agricultural inputs per hectare AIPH (fertilizers, pesticides), for the land fertility LFERT remains nearly constant.
+Food per capita FPC also grows during the 70-year period but at a much
+slower rate than total food F, since the population is also increasing.
 """
 function fig_4(; kwargs...)
     @named ld = Agriculture.land_development()
@@ -138,16 +143,18 @@ function fig_4(; kwargs...)
 end
 
 """
[…]
