+++ paper/paper.bib
@@ -202,3 +202,17 @@
 }
 
 
+
+
+@article{Bezanson2017,
+author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
+title = {Julia: A Fresh Approach to Numerical Computing},
+journal = {SIAM Review},
+volume = {59},
+number = {1},
+pages = {65-98},
+year = {2017},
+doi = {10.1137/141000671},
+url = { https://doi.org/10.1137/141000671},
+}
+
