+++ paper/paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 
 # Summary
 
-\texttt{WorldDynamics.jl} is an open-source Julia package that provides a modern framework for developing, simulating, and investigating Integrated Assessment Models (in short, IAMs). IAMs are numerical models that estimate possible scenarios for the evolution of human society with respect to fundamental aspects such as capital investment, food production, natural resources, population size, and pollution, among others [@Parker2002].  The current version of the package implements the family of IAMs developed by the Club of Rome, which is a non-profit, informal organization of intellectuals and business leaders dedicated to the critical discussion of pressing global issues [@clubofrome2023]. In particular, the library includes Forrester's World1 and World2 models [@Forrester1970; @Forrester1973], the influential World3 model of Meadows et al [@Meadows1974], and its 1994 and 2003 updated versions [@Meadows2004]. These models are typically structured into several sectors, which are, in turn, composed of several differential-algebraic systems of equations (in short, DAEs): this modular structure makes it easy to manipulate the sectors individually and to combine different DAEs into a unique model (see **Figure 1**, which represents the structure of the World3 model). The package benefits from Julia's scientific computing ecosystem, in particular from the efficient solver implementations for various differential equations provided by the \texttt{DifferentialEquations.jl} package [@Rackauckas2017] and from the ability of the \texttt{ModelingToolkit.jl} package to compose differential-algebraic equations [@Ma2021]. \texttt{WorldDynamics.jl} provides also several support functions, which allowed us to convert the original systems of finite-difference equations into DAEs and to easily replicate all the plots included in the above-mentioned references (see, for example, **Figure 2**). It also allows the possibility of easily changing the parameter values and the systems of equations in order to simulate different scenarios. In other words, \texttt{WorldDynamics.jl} facilitates the use of advanced scientific computing approaches for both classical and new models.
+\texttt{WorldDynamics.jl} is an open-source package for the Julia programming language [@Bezanson2017] that provides a modern framework for developing, simulating, and investigating Integrated Assessment Models (in short, IAMs). IAMs are numerical models that estimate possible scenarios for the evolution of human society with respect to fundamental aspects such as capital investment, food production, natural resources, population size, and pollution, among others [@Parker2002].  The current version of the package implements the family of IAMs developed by the Club of Rome, which is a non-profit, informal organization of intellectuals and business leaders dedicated to the critical discussion of pressing global issues [@clubofrome2023]. In particular, the library includes Forrester's World1 and World2 models [@Forrester1970; @Forrester1973], the influential World3 model of Meadows et al [@Meadows1974], and its 1994 and 2003 updated versions [@Meadows2004]. These models are typically structured into several sectors, which are, in turn, composed of several differential-algebraic systems of equations (in short, DAEs): this modular structure makes it easy to manipulate the sectors individually and to combine different DAEs into a unique model (see **Figure 1**, which represents the structure of the World3 model). The package benefits from Julia's scientific computing ecosystem, in particular from the efficient solver implementations for various differential equations provided by the \texttt{DifferentialEquations.jl} package [@Rackauckas2017] and from the ability of the \texttt{ModelingToolkit.jl} package to compose differential-algebraic equations [@Ma2021]. \texttt{WorldDynamics.jl} provides also several support functions, which allowed us to convert the original systems of finite-difference equations into DAEs and to easily replicate all the plots included in the above-mentioned references (see, for example, **Figure 2**). It also allows the possibility of easily changing the parameter values and the systems of equations in order to simulate different scenarios. In other words, \texttt{WorldDynamics.jl} facilitates the use of advanced scientific computing approaches for both classical and new models.
 
 # Statement of need
[…]
