+++ paper/paper.md
@@ -55,7 +55,7 @@ Predicting the evolution of human society is a challenging and pressing question
 ![An example of the reproduction of a figure 7-7, World 3 reference run, included in [@Meadows1974]. The colored curves, which are generated by our package, have been superimposed on the original plots of the book (dotted, dashed, and one-character curves).](images/fig_7_7.png)
 
 # Related packages and implementations
-Besides the official implementation in Stella (which was included in the CD-ROM attached to [@Meadows2004]), which is one of the most popular system dynamics simulation environments, World2 and World3 have been repeatedly implemented in several different programming languages and simulation environments, such as Modelica [@World3Modelica], Simulink [@World3Simulink] and Vensim [@World303Vensim]. Here, we mention some implementations of these models in three popular programming languages. In Python, there are several repositories whose code implements some of the IAMs proposed by the Club of Rome: in our opinion, the most complete description of the World2 and World3 models and of their implementation is given in [@Vanwynsberghe2021]. An implementation of World2 in R is described in [@Mignan2020], while an implementation of the same model in C++ is given in [@Hay2021]. In the Julia ecosystem, there are some IAMs implemented using the Mimi framework [\texttt{Mimi.jl}](https://www.mimiframework.org), such as the one included in [\texttt{MimiPAGE2009.jl}](https://github.com/anthofflab/MimiPAGE2009.jl), a package that implements a model for estimating the social cost of carbon emissions [@Moore2018]. [\texttt{ClimateMARGO.jl}](https://github.com/ClimateMARGO) is an idealized framework for optimizing climate control strategies implementing the MARGO model [@Drake2021].
+Besides the official implementation in Stella (which was included in the CD-ROM attached to [@Meadows2004]), which is one of the most popular system dynamics simulation environments, World2 and World3 have been repeatedly implemented in several different programming languages and simulation environments, such as Modelica [@World3Modelica], Simulink [@World3Simulink] and Vensim [@World3Vensim]. Here, we mention some implementations of these models in three popular programming languages. In Python, there are several repositories whose code implements some of the IAMs proposed by the Club of Rome: in our opinion, the most complete description of the World2 and World3 models and of their implementation is given in [@Vanwynsberghe2021]. An implementation of World2 in R is described in [@Mignan2020], while an implementation of the same model in C++ is given in [@Hay2021]. In the Julia ecosystem, there are some IAMs implemented using the Mimi framework [\texttt{Mimi.jl}](https://www.mimiframework.org), such as the one included in [\texttt{MimiPAGE2009.jl}](https://github.com/anthofflab/MimiPAGE2009.jl), a package that implements a model for estimating the social cost of carbon emissions [@Moore2018]. [\texttt{ClimateMARGO.jl}](https://github.com/ClimateMARGO) is an idealized framework for optimizing climate control strategies implementing the MARGO model [@Drake2021].
 
 
 # Future work
