### 2023-09-17 worlddynamics/WorldDynamics.jl#197 @aurorarossi
We thank the reviewer for his useful comments on our work.

Regarding the contribution, the one showed in the README is outdated and we were aiming at replacing it with a direct reference to a publication on our work (e.g. JOSS, if accepted).
The author @aurorarossi has been included as an author both for her involvement in the development (past and present) of the overall software produced by the [WorldDynamics organization](https://github.com/worlddynamics) and for having been the main author of the submitted paper. While not evident from the line counts of the commit history, @universmile's has contributed less substantially to the software and, after having co-authored the [extended abstract presented at ROADEF 2023](https://hal.science/hal-04008491) (without proceedings), he quit the project development group at the end of October 2022.

Thank you for the remark. We added a Getting Started section to the Documentation as well.

We provided specific pointers to the original figures corresponding to each function in the docstrings of the function itself, which gets exported in the documentation.
Indeed we did not emphasize this. In the Getting Started section that we added (answer to previous comment) we make this more explicit.
We also added the explicit  signature to the docstring of each function producing the figures.

We indeed refrained from providing all original pictures because of copyrights. The book is publicly accessible (with mild limitations) at https://archive.org/details/dynamicsofgrowth0000unse . By referring to this version we were able to verify the correspondence of all plots with the original figures.

The tests currently present are only mocked: they only check whether some figure is produced at the end. Because of the nature of the package, we did not manage, at least as for now, to design meaningful tests. For this reason, we did not deem it worth to document them and automate them. To avoid confusion, we propose to remove the test folder althogether.

We thank the reviewer for the remark. We added references to Julia, and to the Vensim and Stella references. In fact, after carefully checking, the Vensim implementation is not an official one. The only official implementation which should be available is the Stella implementation of World3-03 which should ship in a CD-ROM together with the book by Meadows et al. 2004. We corrected our text accordingly.
We remark that we were not able to retrieve such CD (even by asking to the publisher of the book), and that the Stella software is proprietary and not available on Linux.
