### 2023-09-18 worlddynamics/WorldDynamics.jl#197 @ranocha
> We added a Getting Started section to the Documentation as well.

Please release a new version so that it shows up in the stable docs.
