### 2023-09-18 worlddynamics/WorldDynamics.jl#197 @ranocha
> We do not provide specific guidelines. In the README, we added a statement saying that we adhere to the [general standards of the Julia community](https://julialang.org/community/standards/).

The JOSS review checklist contains the item

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I need something to be able to check this box. You could for example consider setting up a `CONTRIBUTING.md` file with the relevant information.
