+++ CONTRIBUTING.md
@@ -1,0 +1,43 @@
+# Contributing to WorldDynamics.jl 
+
+Thank you for considering contributing to the WorldDynamics.jl open source framework! 
+
+Following these guidelines will help us make the contribution process easy and effective for everyone involved.
+We will do our best to address your issue, evaluate changes, and help you finalize your pull requests.
+
+## Code of Conduct
+
+We take our open source community seriously, and hold ourselves and other contributors to high standards of communication. By participating and contributing to this project, you agree to adhere to the [Julia Community Standards](https://julialang.org/community/standards/).
+
+## Getting Started
+
+Contributions to this repo are made via issues and pull requests (PRs). Here are some general guidelines that apply to both:
+
+- Search for existing issues and PRs before creating your own.
+- We work hard to ensure that issues are handled in a timely manner, but depending on the impact, it may take a while to investigate the root cause. A friendly ping in the comment thread to the submitter or a contributor can help get attention if your issue is being blocked.
+
+
+### Issues
+
+Issues should be used to report problems with the library, request a new feature, or discuss potential changes before a PR is created. 
+
+If you find an issue that addresses the problem you're having, please add your own reproduction information to the existing issue rather than creating a new one. 
+
+### Pull Requests
+
+PRs for our package are always welcome, and can be a quick way to get your fix or improvement scheduled for the next release. In general, PRs should
+
+- Add unit or integration tests for fixed or changed functionality (if a test suite already exists).
+- Include documentation.
+- Be accompanied by a description of the main objective of the PR.
+
+For changes that affect core functionality, or that would require disruptive changes (e.g., a major release), it's best to open an issue first to discuss your proposal. This is not required, but can save time in creating and reviewing changes.
+
+In general, we follow the "fork and pull" Git workflow
+
+1. Fork the repository into your own Github account
+2. Clone the project to your machine
+3. Create a local branch with a short but descriptive name
+4. Commit changes to the branch
+6. Push changes to your fork
+7. Open a PR in our repository.
