+++ README.md
@@ -23,6 +23,10 @@ Here is the output superposed to the original picture:
 
 For how to produce different figures (and much more), please see [the documentation](https://worlddynamics.github.io/WorldDynamics.jl/stable/).
 
+### Contributions guidelines
+
+Contributions are welcome! Please see [CONTRIBUTING.md](CONTRIBUTING.md) for details.
+
 ### How to cite this work 
 
 ```
