+++ test/world3_03.jl
@@ -1,0 +1,2062 @@
+# Vensim solutions
+food = [4.3092e+11
+4.36395e+11
+4.40511e+11
+4.43937e+11
+4.4704e+11
+4.50008e+11
+4.5291e+11
+4.55789e+11
+4.58673e+11
+4.61577e+11
+4.64491e+11
+4.67424e+11
+4.70385e+11
+4.7338e+11
+4.7641e+11
+4.79479e+11
+4.82588e+11
+4.85737e+11
+4.88928e+11
+4.92162e+11
+4.95441e+11
+4.98764e+11
+5.02134e+11
+5.05552e+11
+5.09018e+11
+5.12534e+11
+5.16182e+11
+5.19957e+11
+5.23817e+11
+5.27738e+11
+5.31709e+11
+5.35723e+11
+5.39779e+11
+5.43877e+11
+5.48019e+11
+5.52199e+11
+5.56409e+11
+5.60659e+11
+5.64957e+11
+5.69305e+11
+5.73708e+11
+5.78167e+11
+5.82685e+11
+5.87264e+11
+5.91905e+11
+5.9661e+11
+6.0138e+11
+6.06216e+11
+6.11122e+11
+6.16097e+11
+6.21144e+11
+6.26264e+11
+6.31459e+11
+6.3673e+11
+6.4208e+11
+6.47508e+11
+6.53018e+11
+6.58611e+11
+6.64289e+11
+6.70053e+11
+6.75906e+11
+6.81848e+11
+6.87883e+11
+6.94012e+11
+7.00237e+11
+7.06558e+11
+7.12979e+11
+7.19502e+11
+7.26127e+11
+7.32857e+11
+7.39694e+11
+7.4664e+11
+7.53696e+11
+7.60865e+11
+7.68149e+11
+7.75549e+11
+7.83068e+11
+7.90708e+11
+7.98472e+11
+8.0636e+11
+8.14365e+11
+8.22467e+11
+8.30685e+11
+8.39506e+11
+8.48627e+11
+8.57933e+11
+8.67378e+11
+8.7695e+11
+8.86645e+11
+8.96465e+11
+9.06412e+11
+9.16491e+11
+9.26704e+11
+9.37055e+11
+9.47545e+11
+9.58177e+11
+9.68954e+11
+9.79879e+11
+9.90953e+11
+1.00218e+12
+1.01356e+12
+1.02509e+12
+1.03678e+12
+1.04864e+12
+1.06065e+12
+1.07283e+12
+1.08517e+12
+1.09768e+12
+1.11035e+12
+1.12355e+12
+1.13719e+12
+1.15105e+12
+1.16506e+12
+1.17921e+12
+1.19348e+12
+1.20788e+12
+1.22242e+12
+1.2371e+12
+1.25191e+12
+1.26688e+12
+1.28199e+12
+1.29724e+12
+1.31265e+12
+1.32821e+12
+1.34392e+12
+1.35979e+12
+1.3758e+12
+1.39197e+12
+1.40829e+12
+1.42475e+12
+1.44136e+12
+1.45737e+12
+1.47205e+12
+1.48833e+12
+1.50533e+12
+1.5227e+12
+1.54032e+12
+1.55814e+12
+1.57614e+12
+1.59427e+12
+1.61236e+12
+1.63051e+12
+1.64876e+12
+1.66714e+12
+1.68565e+12
+1.7043e+12
+1.7231e+12
+1.74204e+12
+1.76113e+12
+1.78037e+12
+1.79976e+12
+1.81931e+12
+1.83932e+12
+1.85961e+12
+1.88006e+12
+1.90065e+12
+1.92137e+12
+1.94222e+12
+1.96319e+12
+1.98428e+12
[…]
