+++ test/world3_91/scenarios.jl
@@ -1,5 +1,0 @@
-using WorldDynamics, Test, ModelingToolkit
-
-@testset "World3-91" begin
-    @test isa(World3_91.scenario1(), ModelingToolkit.ODESystem)
-end
