+++ test/runtests.jl
@@ -4,26 +4,6 @@ using WorldDynamics, Test
     include("functions.jl")
 end
 
-@testset verbose=true "WorldDynamics.jl" begin
-
-    @testset verbose=true "Scenarios" begin
-        include("world1/scenarios.jl")
-        include("world1a/scenarios.jl")
-        include("world1b/scenarios.jl")
-        include("world2/scenarios.jl")
-        include("world3/scenarios.jl")
-        include("world3_91/scenarios.jl")
-        include("world3_03/scenarios.jl")
-    end
-
-    @testset verbose=true "Figures" begin
-        include("world1/figures.jl")
-        include("world1a/figures.jl")
-        include("world1b/figures.jl")
-        include("world2/figures.jl")
-        include("world3/figures.jl")
-        include("world3_91/figures.jl")
-        include("world3_03/figures.jl")
-    end
-
+@testset verbose=true "World3_03" begin
+    include("world3_03.jl")
 end
