+++ .github/workflows/Test.yml
@@ -3,7 +3,7 @@ name: Test
 on:
   push:
     branches:
-      - main
+      - new-test
     tags: '*'
   pull_request:
 
