### 2023-09-27 worlddynamics/WorldDynamics.jl#197 @aurorarossi
We have made a new release that includes the new docs with updated Getting Started information, Automatic Test Actions, and Contributing Guidelines in the CONTRIBUTING.md file. We have also updated the README.md accordingly.
Thank you
