### 2023-10-14 #5772 @ranocha
> @ranocha : Are you satisfied with the answers and updated documented here [worlddynamics/WorldDynamics.jl#197](https://github.com/worlddynamics/WorldDynamics.jl/issues/197) ?

👍
