### 2023-12-19 #5772 @fraukewiese
@arbennett : – would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
One review has already been done, but the other reviewer cannot make it. Thus we are looking for a different person and I think that you would be a very suitable expert for that :)
