### 2023-12-19 #5772 @arbennett
@fraukewiese - unfortunately I don't currently have the capacity to help out on this one. The paper/package does look super interesting though!

I'd be happy to recommend other potential reviewers if you'd like - feel free to send me an email: andrbenn@arizona.edu
