### 2024-02-02 #5772 @miguelraz
1 small detail:
In the citations, the `Fiddaman, T. [web.archive.org](http://web.archive.org/...)` link is not accessible via the pdf link (and overruns the margin limit). The formatting is off, but technically not the citation itself AFAIK.
