+++ paper/paper.bib
@@ -205,7 +205,7 @@
   author = {Tom Fiddaman},
   year = {2010},
   month = apr,
-  howpublished = {https://web.archive.org/web/20230205104440/https://metasd.com/2010/04/world3-03/}
+  url = {https://metasd.com/2010/04/world3-03/}
 }
 
 @article{Bezanson2017,
