### 2024-02-04 #5772 @aurorarossi
> 1 small detail: In the citations, the `Fiddaman, T. [web.archive.org](http://web.archive.org/...)` link is not accessible via the pdf link (and overruns the margin limit). The formatting is off, but technically not the citation itself AFAIK.

We have fixed the formatting, thank you for pointing this out.
