### 2024-03-04 #5772 @danielskatz
@aurorarossi - I've suggested a few minor changes in the paper/bib in https://github.com/worlddynamics/WorldDynamics.jl/pull/207 - please merge this, or let me know what you disagree with
